test_quasistatic_sim_gradients with both free and commercial solvers.
Commercial solvers are only used when they are available.
