+++ quasistatic_simulator_cpp/src/quasistatic_simulator.cc
@@ -1590,7 +1590,6 @@ QuasistaticSimulator::CalcDynamics(QuasistaticSimulator *q_sim,
   q_sim->UpdateMbpPositions(q);
   auto tau_ext_dict = q_sim->CalcTauExt({});
   auto q_a_cmd_dict = q_sim->GetQaCmdDictFromVec(u);
-  const auto &sp = q_sim->get_sim_params();
   q_sim->Step(q_a_cmd_dict, tau_ext_dict, sim_params);
   return q_sim->GetMbpPositionsAsVec();
 }
