+++ quasistatic_simulator_cpp/src/test_batch_simulator.cc
@@ -86,13 +86,11 @@ protected:
     const auto idx_o = name_to_idx_map.at(object_name);
 
     VectorXd q_u0(7);
-    q_u0 << 0.96040786, 0.07943188, 0.26694634, 0.00685272, -0.08083068,
-        0.00117524, 0.0711;
+    q_u0 << 1, 0, 0, 0, -0.081, 0.001, 0.071;
 
     VectorXd q_a0(q_sim.num_actuated_dofs());
-    q_a0 << 0.03501504, 0.75276565, 0.74146232, 0.83261002, -0.1438725,
-        0.74696812, 0.61908827, 0.70064279, -0.06922541, 0.78533142, 0.82942863,
-        0.90415436, 0.63256269, 1.02378254, 0.64089555, 0.82444782;
+    q_a0 << 0.035, 0.753, 0.741, 0.833, -0.144, 0.745, 0.619, 0.701, -0.069,
+        0.785, 0.829, 0.904, 0.633, 1.024, 0.641, 0.824;
 
     ModelInstanceIndexToVecMap q0_dict = {{idx_o, q_u0}, {idx_r, q_a0}};
